mbsp/mbsportal-services/mbs-portal-api/src/test/java/com/fanniemae/mbsportal/controller/PingControllerTest.java
/*
 * 
 *  Copyright (c) 2018 Fannie Mae. All rights reserved. Unpublished -- Rights
 *  reserved under the copyright laws of the United States and international
 *  conventions. Use of a copyright notice is precautionary only and does not
 *  imply publication or disclosure. This software contains confidential
 *  information and trade secrets of Fannie Mae. Use, disclosure, or reproduction
 *  is prohibited without the prior written consent of Fannie Mae.
 *  
 */

package com.fanniemae.mbsportal.controller;

import static org.junit.Assert.assertEquals;

import org.junit.Before;
import org.junit.Test;

import com.fanniemae.mbsportal.api.controller.PingController;
import com.fanniemae.mbsportal.utils.exception.MBSBaseException;

/**
 * 
 * @author g8upjv
 *
 */
public class PingControllerTest {

    PingController pingController = new PingController();

    /**
     * Purpose: Creates the productPO data model
     * 
     * @throws Exception
     */
    @Before
    public void setUp() throws Exception {

    }

    /**
     * Purpose: Test for case when media type is empty
     * 
     * @throws MBSBaseException
     * @throws Exception
     */
    @Test
    public void testping() {
        assertEquals(pingController.ping(), "SUCCESS");
    }

}
